+++ lib/widgets/components/tickets/analog_closed_popup.dart
@@ -7,6 +7,8 @@ import 'package:coffeecard/widgets/components/tickets/rounded_button.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
+// TODO: Code needs to be looked
+//       before being used again
 class AnalogClosedPopup extends StatelessWidget {
   const AnalogClosedPopup({Key? key}) : super(key: key);
