+++ lib/widgets/pages/tickets/buy_tickets_page.dart
@@ -37,6 +37,7 @@ class BuyTicketsPage extends StatelessWidget {
                 ),
               );
             } else if (state is BuyTicketsError) {
+              // FIXME handle error
               return const Text('error');
             }
